e2e: proactively/fully close child processes on e2e exit
On process exit for e2e task, use brute force to kill java process
(using tree-kill module on windows an process.exit
on other platforms).

Track closure of child processes more completely and correctly.

Give java up to 5 seconds to shut down before giving up on it.

Closes #669

+++ shared/js/dev-tasks/context.js
@@ -146,14 +146,24 @@ var closeServer = function (server, cb) {
     }
   };
   try {
-    server.on('end', onClosed);
-    server.on('close', onClosed);
-    server.on('exit', onClosed);
+    if (server.on) {
+      // a standard node process
+      server.on('end', onClosed);
+      server.on('close', onClosed);
+      server.on('exit', onClosed);
+    }
     if (server.close) {
+      // a manually managed process of ours
       server.close(onClosed);
     }
     else {
-      server.kill();
+      if (server.kill) {
+        server.kill();
+      }
+      else {
+        // this isn't really an external process
+        cb();
+      }
     }
   }
   catch (err) {
@@ -255,6 +265,10 @@ self = module.exports = {
         // server in activeServers (and thus attempt close them all in parallel)
         _.map(activeServers, function (server, key) {
           return function (cb) {
+            var waitTime = 1000;
+            if (key === 'middle-tier') {
+              waitTime = 5000; //java is slow to exit
+            }
             $.util.log(chalk.green('shutting down ' + key));
             var closed = false;
             setTimeout(function () {
@@ -267,7 +281,7 @@ self = module.exports = {
                 }
                 catch (err) { console.log(err); }
               }
-            }, 1000);
+            }, waitTime);
             closeServer(
               server,
               function (err) {
@@ -376,7 +390,7 @@ self = module.exports = {
         listener.on('error', function (err) {
           console.log(err);
         });
-        self.setActiveServer(port, server);
+        self.setActiveServer(port, listener);
 
         return server;
       }
